+++ desed/desed/__init__.py
@@ -9,3 +9,4 @@ from .generate_synthetic import (
     generate_files_from_jams,
     generate_tsv_from_jams,
 )
+from . import post_process, utils
